QA-112: address review — dedupe mod download, scope login stub
Review feedback on the QA-112 PR:

1. The nxm:// download+forward flow was duplicated from mods.spec.ts.
   Extract it into helpers/modDownload.ts (downloadModViaModManager) and
   use it from both mods.spec.ts (QA-108) and mods-deploy.spec.ts (QA-112).
   The helper carries the cross-tier handling: free "Mod manager download"
   link vs premium "Vortex" button, request-based nxm capture with a
   DOM-scan fallback, and &amp;-decoding before the IPC forward.

2. Drop the global shell.openExternal stub from setupMainWindow (global
   stubs are an anti-pattern). Stub it inside loginToNexus instead, scoped
   to the login flow that triggers the external open, and only once the
   login button is visible so the main-process evaluate doesn't race a
   startup navigation ("Execution context was destroyed").

Verified: QA-108 (free + premium) and QA-112 (free) green; tsc/eslint clean.
